Changed
Faced issue with the rate variable is likely due to it being declared as a private member of the Adjustment class and therefore not directly accessible in the SuperAnnuation class. Since SuperAnnuation inherits from Adjustment, made the rate variable protected in Adjustment to allow derived classes like SuperAnnuation to access it.